+++ h2integrate/converters/hydrogen/steam_methane_reformer.py
@@ -1,9 +1,9 @@
 import numpy as np
 from attrs import field, define
+from openmdao.utils import units
 
 from h2integrate.core.utilities import BaseConfig, merge_shared_inputs
 from h2integrate.core.validators import gt_zero, gte_zero
-from h2integrate.tools.constants import HHV_H2_MJ_PER_KG
 from h2integrate.core.model_baseclasses import (
     CostModelBaseClass,
     CostModelBaseConfig,
@@ -144,6 +144,14 @@ class SteamMethaneReformerPerformanceModel(PerformanceModelBaseClass):
             desc="Unmet hydrogen demand for SMR plant",
         )
 
+        self.add_output(
+            "total_energy_conversion_ratio",
+            val=0.0,
+            shape=1,
+            units="kW*h/kg",
+            desc="Net energy conversion ratio",
+        )
+
     def compute(self, inputs, outputs):
         """
         Compute hydrogen output from natural gas input.
@@ -163,7 +171,7 @@ class SteamMethaneReformerPerformanceModel(PerformanceModelBaseClass):
         # calculate max input and output
         system_capacity_kg_per_hour = inputs["system_capacity"] * (
             1000 / 24
-        )  # plant capacity in kg/h
+        )  # plant capacity in kg/h from tonnes per day
         natural_gas_usage_mmbtu_per_kg = inputs["natural_gas_usage_rate"]
         max_natural_gas_consumption = system_capacity_kg_per_hour * natural_gas_usage_mmbtu_per_kg
         electricity_usage_kWh_per_kg = inputs["electricity_usage_rate"]
@@ -202,16 +210,22 @@ class SteamMethaneReformerPerformanceModel(PerformanceModelBaseClass):
         hydrogen_out = np.minimum.reduce([hydrogen_out_ng, hydrogen_out_elec])
 
         outputs["hydrogen_out"] = hydrogen_out
-        outputs["natural_gas_consumed"] = natural_gas_consumed
-        outputs["electricity_consumed"] = electricity_consumed
+        outputs["natural_gas_consumed"] = hydrogen_out * natural_gas_usage_mmbtu_per_kg
+        outputs["electricity_consumed"] = hydrogen_out * electricity_usage_kWh_per_kg
 
         outputs["rated_hydrogen_production"] = system_capacity_kg_per_hour  # kg/h
 
-        # Convert HHV of H2 from MJ/kg to kW*h/kg
-        hhv_h2_kWh_per_kg = HHV_H2_MJ_PER_KG / (3600.0 * 0.001)
+        # Convert natural gas usage from MMBtu/kg to kW*h/kg
+        energy_conversion_ratio_ng = units.convert_units(
+            inputs["natural_gas_usage_rate"], "MMBtu/kg", "kW*h/kg"
+        )
+        total_energy_conversion_ratio = (
+            energy_conversion_ratio_ng + inputs["electricity_usage_rate"]
+        )
+
         outputs["electrical_rated_hydrogen_production"] = (
-            system_capacity_kg_per_hour * hhv_h2_kWh_per_kg
-        ) / 1000  # convert kg/h to MW using HHV of hydrogen
+            system_capacity_kg_per_hour * total_energy_conversion_ratio
+        ) / 1000  # convert kg/h to MW using energy conversion ratio
 
         max_production = system_capacity_kg_per_hour * len(hydrogen_out) * (self.dt / 3600)
 
@@ -223,6 +237,7 @@ class SteamMethaneReformerPerformanceModel(PerformanceModelBaseClass):
             1 / self.fraction_of_year_simulated
         )
         outputs["unmet_hydrogen_demand"] = inputs["hydrogen_demand"] - hydrogen_out
+        outputs["total_energy_conversion_ratio"] = total_energy_conversion_ratio
 
 
 @define(kw_only=True)
@@ -266,7 +281,6 @@ class SteamMethaneReformerCostModel(CostModelBaseClass):
 
         super().setup()
 
-        # Add inputs specific to the cost model with config values as defaults
         self.add_input(
             "annual_hydrogen_produced",
             val=0.0,
@@ -275,11 +289,19 @@ class SteamMethaneReformerCostModel(CostModelBaseClass):
             desc="Annual hydrogen output from performance model",
         )
         self.add_input(
+            "total_energy_conversion_ratio",
+            val=0.0,
+            units="(kW*h)/kg",
+            desc="Plant electricity usage rate in kWh/kg",
+        )
+        self.add_input(
             "electrical_rated_hydrogen_production",
             val=0.0,
             units="kW",
             desc="Electrical equivalent rated hydrogen production from performance model",
         )
+
+        # Add inputs specific to the cost model with config values as defaults
         self.add_input(
             "unit_capex",
             val=self.config.capex_per_kw,
@@ -319,12 +341,9 @@ class SteamMethaneReformerCostModel(CostModelBaseClass):
 
         hydrogen_out = inputs["annual_hydrogen_produced"]  # kg/year annual profile
 
-        # Convert HHV of H2 from MJ/kg to kW*h/kg
-        hhv_h2_kWh_per_kg = HHV_H2_MJ_PER_KG / (3600.0 * 0.001)
-
-        # Convert the variable O&M from USD/kWh to USD/kg using the HHV of H2
+        # Convert the variable O&M from USD/kWh to USD/kg using the energy conversion ratio
         # USD/kW*h * kW*h/kg = USD/kg
-        variable_opex_per_kg = variable_opex_per_kwh * hhv_h2_kWh_per_kg
+        variable_opex_per_kg = variable_opex_per_kwh * inputs["total_energy_conversion_ratio"][0]
 
         # Calculate variable operating expenses over project life
         variable_om = variable_opex_per_kg * hydrogen_out