+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # Changelog
 
 ## Unreleased
+- Update energy conversion ratio in H2 SMR model [PR 606](https://github.com/NatLabRockies/H2Integrate/pull/606)
 - Update iron models and examples [PR 601](https://github.com/NatLabRockies/H2Integrate/pull/601)
   - Remove outdated iron files
   - Consolidate iron examples into a single main folder
   - Add documentation for Rosner iron DRI and steel EAF models
+<<<<<<< pysam_tidal
 - Add PySAM marine models [PR 607](https://github.com/NatLabRockies/H2Integrate/pull/607)
   - Add tidal resource model
   - Add pysam tidal performance model
   - Add pysam marine hydrokinetic cost model
+=======
+- Breaks out pyomo controller simulation code from base class to individual controllers. [PR 587](https://github.com/NatLabRockies/H2Integrate/pull/587)
+>>>>>>> develop
 
 ## 0.7.1 [March 13, 2026]